+++ packages/components/filters/src/filters.tsx
@@ -1,22 +1,43 @@
+import {
+  type ReactNode,
+  type MouseEvent,
+  type KeyboardEvent,
+  useRef,
+  useState,
+} from 'react';
+import { css } from '@emotion/react';
 import * as Popover from '@radix-ui/react-popover';
+import CollapsibleMotion from '@commercetools-uikit/collapsible-motion';
+import { designTokens } from '@commercetools-uikit/design-system';
 import FlatButton from '@commercetools-uikit/flat-button';
-import SelectInput from '@commercetools-uikit/select-input';
 import {
   CloseBoldIcon,
   FilterIcon,
   PlusBoldIcon,
 } from '@commercetools-uikit/icons';
-import { ReactElement, ReactNode, useState } from 'react';
-import { type TAppliedFilterValue } from './filter-menu/filter-menu';
-import { type TPrimaryButtonProps } from '@commercetools-uikit/primary-button';
-import { type TSelectInputProps } from '@commercetools-uikit/select-input';
+import SelectInput, {
+  TOption,
+  TOptionObject,
+} from '@commercetools-uikit/select-input';
 import Spacings from '@commercetools-uikit/spacings';
-import { designTokens } from '@commercetools-uikit/design-system';
-import { css } from '@emotion/react';
-import messages from './messages';
 import { useIntl } from 'react-intl';
-import { menuStyles } from './filter-menu/filter-menu';
-import CollapsibleMotion from '@commercetools-uikit/collapsible-motion';
+import FilterMenu, {
+  menuStyles,
+  menuBodyStyle,
+  type TAppliedFilterValue,
+} from './filter-menu';
+import messages from './messages';
+import { Badge } from './badge';
+
+interface TAddFilterSelectOption extends TOption {
+  isDisabled?: boolean;
+}
+
+interface TAddFilterOptionGroup extends TOptionObject {
+  options: TAddFilterSelectOption[];
+  label: ReactNode;
+  key: string;
+}
 
 type TAppliedFilter = {
   /**
@@ -29,7 +50,19 @@ type TAppliedFilter = {
   values: TAppliedFilterValue[];
 };
 
-type TFilterConfiguration = {
+export type TFilterConfiguration = {
+  /**
+   * unique identifier for the filter
+   */
+  key: string;
+  /**
+   * formatted message to display the filter name
+   */
+  label: ReactNode;
+  /**
+   * formatted message to display the selected operator value
+   */
+  operatorLabel?: ReactNode;
   /**
    * configuration object for the filter menu.
    */
@@ -41,19 +74,23 @@ type TFilterConfiguration = {
     /**
      * the input in which the user can select which operator should be used for this filter
      */
-    renderOperatorsInput?: () => ReactElement<TSelectInputProps>;
+    renderOperatorsInput?: () => ReactNode;
     /**
      * optional button that allows the user to apply selected filter values
      */
-    renderApplyButton?: () => ReactElement<TPrimaryButtonProps>;
+    renderApplyButton?: () => ReactNode;
     /**
      * controls whether `clear` button in Menu Body Footer is displayed
      */
-    onClearRequest: Function;
+    onClearRequest: (
+      event?: MouseEvent<HTMLButtonElement> | KeyboardEvent<HTMLButtonElement>
+    ) => void;
     /**
      * controls whether `sort` button in Menu Body Header is displayed
      */
-    onSortRequest?: Function;
+    onSortRequest?: (
+      event?: MouseEvent<HTMLButtonElement> | KeyboardEvent<HTMLButtonElement>
+    ) => void;
   };
   /**
    * optional key to group filters together.
@@ -64,17 +101,9 @@ type TFilterConfiguration = {
    */
   isPersistent?: boolean;
   /**
-   * whether or not the filter is disabled
+   * indicates whether the filter is disabled
    */
   isDisabled?: boolean;
-  /**
-   * unique identifier for the filter
-   */
-  key: string;
-  /**
-   * formatted message to display the filter name
-   */
-  label: ReactNode;
 };
 
 type TFilterGroupConfiguration = {
@@ -95,30 +124,43 @@ export type TFiltersProps = {
   appliedFilters: TAppliedFilter[];
   /**
    * configuration for the available filters.
+   *
    */
   filters: TFilterConfiguration[];
   /**
    * optional configuration for filter groups.
+   *
    */
   filterGroups?: TFilterGroupConfiguration[];
+
   /**
-   * controls the `clear all` (added filters) button from the menu list
+   * controls the `clear all` (added filters) button from the menu list, meant to clear the parent application's filter state
    */
-  onClearAllRequest: () => void;
+  onClearAllRequest: (
+    event?: MouseEvent<HTMLButtonElement> | KeyboardEvent<HTMLButtonElement>
+  ) => void;
   /**
-   * optional callback trigger to add a new filter.
+   * optional callback when the add filter button is clicked
    */
-  onAddFilterRequest?: Function;
+  onAddFilterRequest?: (
+    event?: MouseEvent<HTMLButtonElement> | KeyboardEvent<HTMLButtonElement>
+  ) => void;
   /**
    * function to render a search input, selectable from applicable UI Kit components.
    */
-  renderSearchComponent: () => ReactNode;
+  renderSearchComponent: ReactNode;
+  /**
+   * controls whether the filters list is initially open
+   */
+  defaultOpen?: boolean;
 };
 
 const horizontalDividerStyles = css`
+  color: ${designTokens.colorNeutral90};
+  background-color: ${designTokens.colorNeutral90};
   width: 100%;
   height: 1px;
-  border: 1px solid ${designTokens.colorNeutral90};
+  border: 0;
   margin-top: ${designTokens.spacing25};
   margin-bottom: ${designTokens.spacing30};
 `;
@@ -133,64 +175,249 @@ const verticalDividerStyles = css`
 const menuListStyles = css`
   width: 100%;
   display: flex;
+  flex-wrap: wrap;
   align-items: center;
   gap: ${designTokens.spacing20};
 `;
 
-function Filters(props: TFiltersProps) {
+function getFilterOptions(
+  filters: TFilterConfiguration[],
+  filterGroups?: TFilterGroupConfiguration[]
+): (TAddFilterSelectOption | TAddFilterOptionGroup)[] {
+  let filterOptions: (TAddFilterSelectOption | TAddFilterOptionGroup)[] = [];
+  // define option groups
+  if (filterGroups) {
+    filterOptions = filterGroups.map((filterGroup) => ({
+      label: filterGroup.label,
+      key: filterGroup.key,
+      options: [],
+    }));
+  }
+  return filters.reduce((filterOptions, filter) => {
+    const formattedOption = {
+      value: filter.key,
+      label: filter.label,
+      isDisabled: filter.isDisabled,
+    };
+    //if theres a groupkey, filterGroups, and the groupKey matches a filterGroup, add option to its group
+    if (filter.groupKey && filterGroups) {
+      const optionGroup = filterOptions.find(
+        (option) => 'key' in option && option.key === filter.groupKey
+      );
+      if (optionGroup && 'options' in optionGroup) {
+        optionGroup.options.push(formattedOption);
+        return filterOptions;
+      }
+    }
+    // otherwise add option directly
+    return [formattedOption, ...filterOptions];
+  }, filterOptions);
+}
+
+function Filters({
+  appliedFilters,
+  filters,
+  filterGroups,
+  onClearAllRequest,
+  onAddFilterRequest,
+  renderSearchComponent,
+  defaultOpen = false,
+}: TFiltersProps) {
   const intl = useIntl();
-  const [showAddFilters, setShowAddFilters] = useState(false);
+  const [showFilterControls, setShowFilterControls] = useState(defaultOpen);
 
   const handleFiltersClick = () => {
-    setShowAddFilters((currState) => !currState);
+    setShowFilterControls((currState) => !currState);
   };
+  /**
+   * persisted filters: always visible
+   * applied filters: filters for which values have been selected
+   * visible filters = persisted + applied
+   *
+   * visibleFiltersFromProps = filters visible based on props
+   * localVisibleFilters = filters actually visible in component currently
+   */
+  const persistedFilterKeys = filters
+    .filter(({ isPersistent }) => isPersistent)
+    .map((filter) => filter.key);
+
+  const persistedFiltersRef = useRef<string[]>(persistedFilterKeys);
+
+  const appliedFilterKeys = appliedFilters.map(({ filterKey }) => filterKey);
+
+  // applied filters must have corresponding filter in `props.filters`,
+  const visibleFiltersFromProps = filters
+    .filter(({ key, isPersistent }) => {
+      const isVisible =
+        Boolean(isPersistent) || appliedFilterKeys.includes(key);
+      return isVisible;
+    })
+    // persistent filters should be first in filter list
+    .sort(({ isPersistent }) => (isPersistent ? -1 : 1));
+
+  // set initial state as visibleFiltersFromProps
+  const [localVisibleFilters, setLocalVisibleFilters] = useState<string[]>(
+    visibleFiltersFromProps.map(({ key }) => key)
+  );
+
+  //update localVisibleFilters if persisted filter count changes
+  if (persistedFiltersRef.current.length !== persistedFilterKeys.length) {
+    setLocalVisibleFilters(visibleFiltersFromProps.map(({ key }) => key));
+    persistedFiltersRef.current = persistedFilterKeys;
+  }
+
+  const removeFilter = (filterKey: string) =>
+    setLocalVisibleFilters((currentVisibleFilters) =>
+      currentVisibleFilters.filter(
+        (visibleFilterKey) => visibleFilterKey !== filterKey
+      )
+    );
 
   return (
     <>
       <Spacings.Inline scale="m" alignItems="center">
         <div css={{ maxWidth: `${designTokens.constraint16}` }}>
-          {props.renderSearchComponent()}
+          {renderSearchComponent}
         </div>
-        <FlatButton
-          data-testid="filters-button"
-          label={intl.formatMessage(messages.filtersButtonLabel)}
-          icon={<FilterIcon />}
-          onClick={handleFiltersClick}
-        />
+        <Spacings.Inline scale="s" alignItems="center">
+          <FlatButton
+            label={intl.formatMessage(messages.filtersButtonLabel)}
+            icon={<FilterIcon />}
+            onClick={handleFiltersClick}
+          />
+          {appliedFilters.length > 1 && !showFilterControls && (
+            <Badge
+              id={'uikit-filters-selected-filter-count'}
+              label={`${appliedFilters.length}`}
+            />
+          )}
+        </Spacings.Inline>
       </Spacings.Inline>
-      <div css={horizontalDividerStyles} />
-
+      <hr css={horizontalDividerStyles} />
       <CollapsibleMotion
-        isClosed={!showAddFilters}
-        onToggle={() => setShowAddFilters(!showAddFilters)}
+        isClosed={!showFilterControls}
+        onToggle={() => setShowFilterControls(!showFilterControls)}
       >
         {({ registerContentNode, containerStyles }) => (
           <div style={containerStyles}>
-            <div ref={registerContentNode} css={menuListStyles}>
+            <div
+              ref={registerContentNode}
+              css={menuListStyles}
+              aria-live="polite"
+            >
+              {localVisibleFilters.map((activeFilter) => {
+                const activeFilterConfig = filters.find(
+                  (filter) => filter.key === activeFilter
+                )!;
+                return (
+                  <FilterMenu
+                    key={activeFilterConfig.key}
+                    filterKey={activeFilterConfig.key}
+                    label={activeFilterConfig.label}
+                    operatorLabel={activeFilterConfig.operatorLabel}
+                    isPersistent={activeFilterConfig.isPersistent}
+                    isDisabled={activeFilterConfig.isDisabled}
+                    renderMenuBody={
+                      activeFilterConfig.filterMenuConfiguration.renderMenuBody
+                    }
+                    renderOperatorsInput={
+                      activeFilterConfig.filterMenuConfiguration
+                        .renderOperatorsInput
+                    }
+                    renderApplyButton={
+                      activeFilterConfig.filterMenuConfiguration
+                        .renderApplyButton
+                    }
+                    appliedFilterValues={
+                      appliedFilters.find(
+                        (appliedFilter) =>
+                          activeFilter === appliedFilter.filterKey
+                      )?.values
+                    }
+                    onClearRequest={
+                      activeFilterConfig.filterMenuConfiguration.onClearRequest
+                    }
+                    onRemoveRequest={(e) => {
+                      removeFilter(activeFilter);
+                      activeFilterConfig.filterMenuConfiguration.onClearRequest(
+                        e
+                      );
+                    }}
+                    onSortRequest={
+                      activeFilterConfig.filterMenuConfiguration.onSortRequest
+                    }
+                    defaultOpen={
+                      activeFilterConfig.isPersistent ||
+                      localVisibleFilters.length ===
+                        visibleFiltersFromProps.length
+                        ? false
+                        : true
+                    }
+                  />
+                );
+              })}
               <Popover.Root>
                 <Popover.Trigger asChild>
                   <div css={{ display: 'inline-flex' }}>
                     <FlatButton
-                      data-testid="add-filter-button"
                       label={intl.formatMessage(messages.addFilterButtonLabel)}
                       icon={<PlusBoldIcon />}
-                      onClick={() => setShowAddFilters(true)}
+                      onClick={(e) => {
+                        if (onAddFilterRequest) {
+                          onAddFilterRequest(e);
+                        }
+                        setShowFilterControls(true);
+                      }}
                     />
                   </div>
                 </Popover.Trigger>
                 <Popover.Portal>
-                  <Popover.Content side="bottom" align="start" css={menuStyles}>
-                    <SelectInput />
+                  <Popover.Content
+                    side="bottom"
+                    align="start"
+                    css={[menuStyles, menuBodyStyle]}
+                  >
+                    <SelectInput
+                      id="ui-kit-add-filters-select"
+                      name="select filters"
+                      aria-label="select filters"
+                      appearance="filter"
+                      options={
+                        getFilterOptions(filters, filterGroups) as TOption[]
+                      }
+                      onChange={(e) => {
+                        setLocalVisibleFilters(
+                          Array.prototype.concat(
+                            e.target.value ? e.target.value : []
+                          )
+                        );
+                      }}
+                      value={localVisibleFilters}
+                      isMulti={true}
+                      // @ts-ignore
+                      isOptionDisabled={(option: TAddFilterSelectOption) =>
+                        option.isDisabled
+                      }
+                    />
                   </Popover.Content>
                 </Popover.Portal>
               </Popover.Root>
-              <div css={verticalDividerStyles} />
-              <FlatButton
-                icon={<CloseBoldIcon />}
-                label={intl.formatMessage(messages.clearAllFiltersButtonLabel)}
-                onClick={props.onClearAllRequest}
-                tone="secondary"
-              />
+              {appliedFilters.length > 1 && (
+                <>
+                  <div css={verticalDividerStyles} />
+                  <FlatButton
+                    icon={<CloseBoldIcon />}
+                    label={intl.formatMessage(
+                      messages.clearAllFiltersButtonLabel
+                    )}
+                    onClick={(e) => {
+                      onClearAllRequest(e);
+                      setLocalVisibleFilters(persistedFilterKeys);
+                    }}
+                    tone="secondary"
+                  />
+                </>
+              )}
             </div>
           </div>
         )}