+++ packages/components/fields/date-field/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @commercetools-uikit/date-field
 
+## 19.16.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/date-input@19.16.0
+  - @commercetools-uikit/design-system@19.16.0
+  - @commercetools-uikit/constraints@19.16.0
+  - @commercetools-uikit/field-errors@19.16.0
+  - @commercetools-uikit/field-label@19.16.0
+  - @commercetools-uikit/field-warnings@19.16.0
+  - @commercetools-uikit/utils@19.16.0
+  - @commercetools-uikit/spacings@19.16.0
+
 ## 19.15.0
 
 ### Patch Changes