+++ packages/components/fields/search-select-field/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @commercetools-uikit/search-select-field
 
+## 19.16.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/search-select-input@19.16.0
+  - @commercetools-uikit/design-system@19.16.0
+  - @commercetools-uikit/constraints@19.16.0
+  - @commercetools-uikit/field-errors@19.16.0
+  - @commercetools-uikit/field-label@19.16.0
+  - @commercetools-uikit/field-warnings@19.16.0
+  - @commercetools-uikit/hooks@19.16.0
+  - @commercetools-uikit/utils@19.16.0
+  - @commercetools-uikit/spacings@19.16.0
+
 ## 19.15.0
 
 ### Patch Changes