+++ packages/components/inputs/localized-rich-text-input/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @commercetools-uikit/localized-rich-text-input
 
+## 19.16.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/design-system@19.16.0
+  - @commercetools-uikit/flat-button@19.16.0
+  - @commercetools-uikit/collapsible-motion@19.16.0
+  - @commercetools-uikit/constraints@19.16.0
+  - @commercetools-uikit/icons@19.16.0
+  - @commercetools-uikit/input-utils@19.16.0
+  - @commercetools-uikit/rich-text-utils@19.16.0
+  - @commercetools-uikit/messages@19.16.0
+  - @commercetools-uikit/spacings-inline@19.16.0
+  - @commercetools-uikit/spacings-stack@19.16.0
+  - @commercetools-uikit/text@19.16.0
+  - @commercetools-uikit/tooltip@19.16.0
+  - @commercetools-uikit/hooks@19.16.0
+  - @commercetools-uikit/localized-utils@19.16.0
+  - @commercetools-uikit/utils@19.16.0
+
 ## 19.15.0
 
 ### Patch Changes