+++ packages/components/inputs/time-input/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @commercetools-uikit/time-input
 
+## 19.16.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/design-system@19.16.0
+  - @commercetools-uikit/accessible-button@19.16.0
+  - @commercetools-uikit/constraints@19.16.0
+  - @commercetools-uikit/icons@19.16.0
+  - @commercetools-uikit/input-utils@19.16.0
+  - @commercetools-uikit/spacings-inline@19.16.0
+  - @commercetools-uikit/hooks@19.16.0
+  - @commercetools-uikit/utils@19.16.0
+
 ## 19.15.0
 
 ### Patch Changes